@@ -1,35 +1,25 @@
 name: periodic_table_app
-description: A Flutter app for learning about the Periodic Table.
-version: 1.0.0
+description: "A new Flutter project."
+
+publish_to: 'none'
+
+version: 1.0.0+1
+
 environment:
-  sdk: '>=2.18.0 <3.0.0'
+  sdk: ^3.8.1
 
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^1.0.5
-  get_it: ^7.6.0
-  flutter_riverpod: ^2.0.0
-  http: ^0.13.5
-  shared_preferences: ^2.2.0
-  flutter_local_notifications: ^12.0.0
-  path: ^1.8.3
-  sqflite: ^2.0.3
-  flutter_spinkit: ^5.2.0
-  url_launcher: ^6.1.7
-  flutter_secure_storage: ^7.0.0
-  connectivity_plus: ^6.0.0
-  fl_chart: ^0.66.2
+  get_it: ^7.2.0
+  cupertino_icons: ^1.0.8
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  mockito: ^5.3.2
+  flutter_lints: ^5.0.0
 
 flutter:
   uses-material-design: true
   assets:
-    - assets/elements.json
-    - assets/quiz_questions.json
-    - assets/learning_paths.json
     - assets/images/logo.jpg
