@@ -1,9 +1,11 @@
 import 'package:flutter/material.dart';
 
 class AppColors {
-  static const Color primaryColor = Color(0xFF1976D2); // Blue for metals
-  static const Color secondaryColor = Color(0xFFD32F2F); // Red for non-metals
-  static const Color accentColor = Color(0xFF388E3C); // Green for noble gases
-  static const Color backgroundColor = Color(0xFFF5F5F5);
-  static const Color textColor = Color(0xFF212121);
+  static const Color primaryColor = Colors.blue;
+  static const Color secondaryColor = Colors.amber;
+  static const Color backgroundColor = Colors.white;
+  static const Color textColor = Colors.black87;
+  static const Color metalColor = Colors.blueGrey;
+  static const Color nonMetalColor = Colors.redAccent;
+  static const Color nobleGasColor = Colors.green;
 }
